Fixing a bug with JjMedia.

+++ src/cake/app/plugins/jj_media/models/behaviors/stored_file_holder.php
@@ -103,7 +103,7 @@ class StoredFileHolderBehavior extends ModelBehavior
 			$SfilStoredFile =& $this->getFileModel($Model);
 			
 			foreach ($this->settings[$Model->alias]['keys'] as $key)
-				if ($Model->data[$Model->alias][$key] != $this->runtime[$Model->alias]['ids_bkp'][$key])
+				if (isset($Model->data[$Model->alias][$key]) && $Model->data[$Model->alias][$key] != $this->runtime[$Model->alias]['ids_bkp'][$key])
 					$SfilStoredFile->delete($this->runtime[$Model->alias]['ids_bkp'][$key]);
 		}
 	}
